use separate module name for extension.schema.{ts,json} files to avoid confusion
Using separate module names makes build configuration simpler (e.g., it removes the need to specify the preferred ordering of module file extensions in Babel, Jest, and Webpack).
