+++ shared/src/extensions/extensionManifest.ts
@@ -1,5 +1,5 @@
 import { isPlainObject } from 'lodash'
-import { ExtensionManifest as ExtensionManifestSchema } from '../schema/extension.schema'
+import { ExtensionManifest as ExtensionManifestSchema } from '../schema/extensionSchema'
 import { ErrorLike, isErrorLike } from '../util/errors'
 import { parseJSONCOrError } from '../util/jsonc'
