+++ shared/src/schema/extensionSchema.ts
