+++ web/src/extensions/ExtensionCard.tsx
@@ -6,7 +6,7 @@ import { Path } from '../../../shared/src/components/Path'
 import { ConfiguredRegistryExtension, isExtensionEnabled } from '../../../shared/src/extensions/extension'
 import * as GQL from '../../../shared/src/graphql/schema'
 import { PlatformContextProps } from '../../../shared/src/platform/context'
-import { ExtensionManifest } from '../../../shared/src/schema/extension.schema'
+import { ExtensionManifest } from '../../../shared/src/schema/extensionSchema'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { isErrorLike } from '../../../shared/src/util/errors'
 import { isExtensionAdded } from './extension/extension'
