+++ web/src/extensions/extension/RegistryExtensionOverviewPage.tsx
@@ -5,7 +5,7 @@ import GithubCircleIcon from 'mdi-react/GithubCircleIcon'
 import * as React from 'react'
 import { Link } from 'react-router-dom'
 import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
-import { ExtensionCategory } from '../../../../shared/src/schema/extension.schema'
+import { ExtensionCategory } from '../../../../shared/src/schema/extensionSchema'
 import { isErrorLike } from '../../../../shared/src/util/errors'
 import { isDefined } from '../../../../shared/src/util/types'
 import { PageTitle } from '../../components/PageTitle'
